remote auto-generated (metadata) file AssemblyInfo.cs from repository
Under the default configuration, this file is generated automatically during a build.
When present in the repository, a `dotnet build` straight out of a `git clone` will fail
due to duplicate metadata.
If it's desired to keep an explicit AssemblyInfo.cs, then the project file should
specify `<GenerateAssemblyInfo>false</GenerateAssemblyInfo>`.